Created Fontgenerator Project
The Fontgenerator Project is a side- project for the creation of
bitmap-fonts via commandline.
The reason for that is the fact that XNA converts TrueType (somehow) to
Bitmap internally. I will do the same, since SpriteFont is not working
without the usage of the contentpipeline.

Added Project "FontGenerator"
Added FontTemplate for easier setup
Added Converter to mange the conversion