Pipeline Update
Added Settings, a static class for internal management (see below)

Changed Renderers and Transform to use Settings' Unit-property
Changed renderers to use a Material defined by the baserenderer (forces
renderers to have a callable material)
Changed Asset to make Name and Filename public (could be called a
prerelease bugfix)
Changed Material

Mayor changes in this update are the aquire and unload methods for
assets. To make the internal assetmanagement simple, I added these
methods, as well as a list of assets. Assets can be aquired and loaded.
If they are already in the list (defined by their name and filename)
they will just be returned. The assets can be unloaded by calling the
unload-method.