Pathfinding Update
Yea, I wanted to refine my code a bit, so I started implementing it.

Added comments to the Node.cs and made the Node class and the enum
public (because of inconsistency in the accessability)
Updated AStar and the classes BoolAStar and ByteAStar. Implemented
Mapgeneration and added Fields and Properties to AStar.