Added Resourcefiles
These to xml files are used by the engine to determine what files are
definitly required (DefaultResources.xml) so the engine has a fallback
when the desired assets are unavailable to execute and what directories
should be cached (Loccations.xml). When loading a resource the folder
will be checked by the engine depending on the resource.