Assetloading Update
This update contains a lot of updates (Updateception)
The mayor one is the fact that the engine wants to initialize requested
files as default, which it won't find unfortunatly (I will provide them
tomorrow).

The problem is that this update goes through a lot of classes since the
initialization by xml is a bad idea, I had to write a static
Create-Method in all the subclasses of Asset. The reason for that is the
easier intialization (as you expected). If there is a better method,
just make a pull request.

Updated Image, ImageMap, Level, Material, Shader, AudioFile and Font
with Create-Method.
Updated Bootstrap with "Forced Loadings" in the constructor.
Fixed Bug in Debug (forgot params in the methods)
Updated Settings: It reads locations and the default assets from the xml
in resources.
It saves the type in the dictionaries with its location. This helps the
engine initializing the asset. You can get the path to the type using a
method. See Documentation for more information.