Asset Update
Updated the Assetclass (and derived):
The AquireAsset-Method will throw an error if loading failed for several
reasons.

Added try-catch to ALL assetloaders to keep the engine running.

Removed the hardcoded-box (although it was awesome) in EngineSpriteBatch
as well as the part it was reffering to.

Added LogDepth to the Debugger.