Inputmanager
Added InputManager
A class used by the whole Engine to get the inputkeys by names. You can
store a name with a corresponding key (e.g. "Left" with Keys.Left). This
can be used to change the keymapping in the engine. In the next update,
a methodgroup will be added to access these keys by name to check if
they are hit. (maybe I can make a workaround with events, but I am not
sure about this yet).

Minor Fix in Textbox to overwrite existing property.