Major Font-Update
This update adds support for bitmap-fonts, e.g. fonts created in a
drawing software.

GameEngine.Font:
-Added CharElement
-Added BitmapEncoder
-Minor filechanges

The CharElement represents a char and the rectangle from which the
BitmapEncoder should get the image-information. The gef-filesystem
itself is not changed.

FontGenerator:
-Added tab in MainWindow:
The new tab is made for the bitmapfonts.
1. You load an image
2. You choose wether the characterdimensions in the font should be
unified or individual
3. The unified setting requires the characters to be given. It will
start from the upper left and work its way to the lower right of the
image.
4. The fontname with which the font will be exported

The save-and load of individual fontsets is also possible using the
buttons on the lower right.

-Advanced Editor:
The Advanced Editor can be called in the individual setting for the
bitmapfonts and is used to assign chars to areas on the image.

It is possible to add new elements, assign chars and the locations.
Next to the list is a preview-window that shows the active selection.
The backgroundcolor can be changed using the boxes on its left.